% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_study.R
\name{xsdid.mc}
\alias{xsdid.mc}
\title{Useful for monte-Carlo study about different variants of Synth-DID and DID}
\usage{
xsdid.mc(
  N = 100,
  T = 200,
  tau = 50,
  beta.tt = 50,
  beta.x = 50,
  beta.xc = 0,
  beta.treat = 50,
  x.tt.load = 0,
  xc.x.load = 0,
  x.treat.load.mean = 3/4,
  x.contr.load.mean = 1/4,
  x.load.alpha.beta = 2,
  xc.treat.load.mean = 3/4,
  xc.contr.load.mean = 1/4,
  xc.load.alpha.beta = 2,
  ar = 0.97,
  ar.tt = ar,
  ar.x = ar,
  ar.xc = ar,
  return.data = FALSE
)
}
\description{
For details look at the code. For an illustration see the vignette.
}
