% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{xsdid_estimate}
\alias{xsdid_estimate}
\title{A shortcut function to directly perform SDID estimation with covariates from a panel data set}
\usage{
xsdid_estimate(
  panel,
  unit = 1,
  time = 2,
  outcome = 3,
  treatment = 4,
  x,
  x.rows = NULL
)
}
\arguments{
\item{panel}{A data frame with columns consisting of units, time, outcome, and treatment indicator. It should be a balanced panel and not contain any NA.}

\item{unit}{The column number/name corresponding to the unit identifier. Default is 1.}

\item{time}{The column number/name corresponding to the time identifier. Default is 2.}

\item{outcome}{The column number/name corresponding to the outcome identifier. Default is 3.}

\item{treatment}{The column number/name corresponding to the treatment status. Default is 4.}

\item{x}{The column numbers/names of all additional control variables}

\item{x.rows}{To estimate the effect of x on y, we use by default all rows in which no treatment takes places. The argument x.rows allows to specify these rows manually. E.g. you could only take all rows before the first treatment starts (i.e. compared to the default exclude the rows of the control group during the treatment period).}
}
\value{
The resulting object from the call to synthdid_estimate.
}
\description{
Calls adjust.outcome.for.y, panel.matrices and then synthdid_estimate. Has similar syntax to
xsdid_se_bootstrap.
}
\examples{
\dontrun{
 dat = xsdid.mc(N=20, T=20,return.data = TRUE)
 xsdid_estimate(dat, unit="i",time = "t",outcome = "y",treatment = "treat_exp",x = "x")
 xsdid_se_bootstrap(dat, unit="i",time = "t",outcome = "y",treatment = "treat_exp",x = "x", B=100)$se
}
}
